requirements.txt
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.1.3
fastapi==0.104.0
ffmpy==0.3.1
filelock==3.12.4
fonttools==4.43.1
fsspec==2023.9.2
gradio==3.50.1
gradio_client==0.6.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.26.1
orjson==3.9.9
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
websockets==11.0.3
zipp==3.17.0
